refactor: centralize state journal and fs primitives
