+++ src/state/mod.rs
@@ -1,10 +1,8 @@
+pub(crate) mod journal;
 mod lock;
 mod ops;
 
-pub use ops::{
-    OpsAuditOperation, remove_path_if_exists, summarize_history_body,
-    synthesize_snapshot_raw_from_segment_bodies,
-};
+pub use ops::OpsAuditOperation;
 
 use lock::{LockMetadata, acquire_lock_coordinator, lock_file_matches_owner, try_reap_stale_lock};
 
@@ -34,6 +32,7 @@ pub struct AppContext {
     pub pending_ops_file: PathBuf,
     pub pending_ops_history_dir: PathBuf,
     pub pending_ops_snapshot_file: PathBuf,
+    pub command_events_file: PathBuf,
     in_proc: InProcMap,
 }
 
@@ -219,6 +218,7 @@ impl AppContext {
         let pending_ops_file = state_dir.join("pending_ops.jsonl");
         let pending_ops_history_dir = state_dir.join("pending_ops_history");
         let pending_ops_snapshot_file = state_dir.join("pending_ops_snapshot.json");
+        let command_events_file = state_dir.join("events/commands.jsonl");
 
         Ok(Self {
             root,
@@ -228,6 +228,7 @@ impl AppContext {
             pending_ops_file,
             pending_ops_history_dir,
             pending_ops_snapshot_file,
+            command_events_file,
             in_proc: Arc::new(Mutex::new(HashMap::new())),
         })
     }
@@ -458,29 +459,6 @@ fn ensure_file_with_contents(path: &Path, contents: &str) -> Result<()> {
     write_atomic(path, contents).with_context(|| format!("failed to initialize {}", path.display()))
 }
 
-fn append_lines(path: &Path, lines: &[String]) -> Result<()> {
-    if lines.is_empty() {
-        return Ok(());
-    }
-    let parent = path
-        .parent()
-        .context("cannot append file without parent directory")?;
-    fs::create_dir_all(parent)
-        .with_context(|| format!("failed to create parent directory {}", parent.display()))?;
-    let mut file = OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(path)
-        .with_context(|| format!("failed to open {}", path.display()))?;
-    for line in lines {
-        writeln!(file, "{}", line)
-            .with_context(|| format!("failed to append {}", path.display()))?;
-    }
-    file.sync_all()
-        .with_context(|| format!("failed to sync {}", path.display()))?;
-    Ok(())
-}
-
 fn write_history_segment_if_missing(path: &Path, raw: &str) -> Result<()> {
     if raw.is_empty() {
         return Ok(());
@@ -517,13 +495,6 @@ fn write_history_segment_if_missing(path: &Path, raw: &str) -> Result<()> {
     };
     write_atomic(path, &normalized)
         .with_context(|| format!("failed to write {}", path.display()))?;
-    Ok(())
-}
-
-fn maybe_fault_inject(tag: &str) -> Result<()> {
-    if std::env::var("LOOM_FAULT_INJECT").ok().as_deref() == Some(tag) {
-        return Err(anyhow::anyhow!("fault injected at {}", tag));
-    }
     Ok(())
 }
